+++ Properties/AssemblyInfo.cs
@@ -4,5 +4,5 @@ using System.Reflection;
 [assembly: AssemblyDescription("YAPYAP save protection and configurable failed-extraction recovery")]
 [assembly: AssemblyCompany("SaveGuard")]
 [assembly: AssemblyProduct("SaveGuard")]
-[assembly: AssemblyVersion("0.1.0.0")]
-[assembly: AssemblyFileVersion("0.1.0.0")]
+[assembly: AssemblyVersion("0.1.4.0")]
+[assembly: AssemblyFileVersion("0.1.4.0")]